Update 25/10/2014
Added RapidXML to the project as most of the data files will be in
format and there is no point writing an XML parser when there is a
perfectly good free one.
Added the initial sections of the memory map, allocators and pools.
Added headers for modules as we build the application.  Also added a new
utility section to the kernel for things like min/max and ispowertwo
functions.